Moved JitmTracker.kt to jitm package
